[BE] Ref: 단일 책임 원칙에 따라 서비스계층 리팩토링
