+++ src/main/java/comment/bean/Comment.java
@@ -3,7 +3,6 @@ package comment.bean;
 import article.bean.Article;
 import jakarta.persistence.*;
 import lombok.Data;
-import lombok.ToString;
 import member.bean.Member;
 
 import java.time.LocalDateTime;
@@ -31,7 +30,7 @@ public class Comment {
     @JoinColumn(name = "parent_id")
     private Comment parent; // 부모 댓글 참조
 
-    @OneToMany(mappedBy = "parent", cascade = CascadeType.ALL, orphanRemoval = true, fetch = FetchType.EAGER)
+    @OneToMany(mappedBy = "parent", cascade = CascadeType.REMOVE, orphanRemoval = true, fetch = FetchType.LAZY)
     private List<Comment> children = new ArrayList<>(); // 대댓글 리스트
 
     @Column(length = 5000, nullable = false)
@@ -39,7 +38,7 @@ public class Comment {
 
     @Column(name = "comment_status")
     @Enumerated(EnumType.STRING)
-    private CommentStatus commentStatus = CommentStatus.DEFAULT; // 댓글 상태
+    private CommentStatus commentStatus; // 댓글 상태
 
     @Column(name = "write_date", nullable = false)
     private LocalDateTime writeDate; //댓글 작성 시간
@@ -47,6 +46,9 @@ public class Comment {
     @PrePersist
     public void prePersist() {
         this.writeDate = LocalDateTime.now();
+        if (this.commentStatus == null) {
+            this.commentStatus = CommentStatus.DEFAULT;
+        }
     } // 자동으로 작성 시간 입력
 
 }
